requirements.txt
absl-py==1.4.0
accelerate==0.21.0
addict==2.4.0
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2022.12.7
charset-normalizer==2.1.1
clearml==1.12.2
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
cmake==3.25.0
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
-e git+https://github.com/KaiyangZhou/Dassl.pytorch.git@c61a1b570ac6333bd50fb5ae06aea59002fb20bb#egg=dassl
debugpy==1.6.7
decorator==5.1.1
easydict==1.10
entrypoints==0.3
executing==1.2.0
filelock==3.9.0
flake8==3.7.9
fonttools==4.41.0
fsspec==2023.6.0
ftfy==6.1.1
furl==2.1.3
future==0.18.3
gdown==4.7.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grad-cam==1.4.8
grpcio==1.56.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
ipykernel==6.24.0
ipython==8.14.0
isort==4.3.21
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
lit==15.0.7
littleutils==0.2.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.6.1
mpmath==1.2.1
nest-asyncio==1.5.6
networkx==3.0
numpy==1.24.0
oauthlib==3.2.2
ogb==1.3.6
opencv-python==4.8.1.78
orderedmultidict==1.0.1
outdated==0.2.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathlib2==2.3.7.post1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.9.1
prettytable==3.8.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.15.1
PyJWT==2.4.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
randaugment==1.0.2
referencing==0.30.2
regex==2023.6.3
requests==2.28.1
requests-oauthlib==1.3.1
rpds-py==0.9.2
rsa==4.9
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
soupsieve==2.4.1
stack-data==0.6.2
sympy==1.11.1
tabulate==0.9.0
tb-nightly==2.14.0a20230714
tensorboard-data-server==0.7.1
termcolor==2.3.0
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
triton==2.0.0
ttach==0.0.3
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
wcwidth==0.2.6
Werkzeug==2.3.6
wilds==1.2.2
yacs==0.1.8
yapf==0.29.0
zipp==3.16.1
